% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{austin}
\alias{austin}
\title{Austin rides}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 166 rows and 61 columns.}
\usage{
austin
}
\keyword{datasets}
