% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{austin5w}
\alias{austin5w}
\title{Austin rides}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2011 rows and 61 columns.}
\usage{
austin5w
}
\keyword{datasets}
