% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cluster_predictive}
\alias{cluster_predictive}
\title{Computes the cluster predictive distribution of a point}
\description{
Computes the predictive distribution of each cluster for a new point \eqn{x}.
The updated hyperparameters are
}
\details{
The notation differs slightly from the paper
}
\references{
Matthew A. Taddy (2012) Autoregressive Mixture Models for Dynamic Spatial Poisson Processes:
 Application to Tracking Intensity of Violent Crime, Journal of the American Statistical Association,
 105:492, 1403-1417, DOI: 10.1198/jasa.2010.ap09655
}
