% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DDPN}
\alias{DDPN}
\title{DDPN class}
\description{
This class is used inside C++ only and can't be called from R.
It carries the a collection of Dynamic Particles for fitting DP Dynamic Mixture with particle filters,
and the information for the prior.
\itemize{
  \item \eqn{N}: Number of particles
  \item \code{hp}: and object of class DDPNHyperParam
  \item \code{particle_list}: a vector of size \eqn{N} of objects of class DynamicParticle
}
}
\references{
Matthew A. Taddy (2012) Autoregressive Mixture Models for Dynamic Spatial Poisson Processes:
 Application to Tracking Intensity of Violent Crime, Journal of the American Statistical Association, 
 105:492, 1403-1417, DOI: 10.1198/jasa.2010.ap09655
}
