% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DDPNHyperParam}
\alias{DDPNHyperParam}
\title{DDPNHyperParam class}
\description{
This class is used inside C++ only and can't be called from R.
It carries the prior distribution for the parameters for Dirichlet Process Mixture
\deqn{x_i \mid (\mu, \Sigma) \sim N(\mu, \Sigma)}
\deqn{(\mu, \Sigma) \sim \text{DP}(\alpha, \text{NormalInvWishart}(\lambda, \kappa, \Omega, \nu))}
The stick-breaking weights follow a \eqn{BAR(\rho)} process. The hyperparameters are
\itemize{
  \item \eqn{\alpha}: The weights at time \eqn{\tau} have prior 
  \item \eqn{\lambda}: mean, \eqn{\mu \sim N(\lambda, \Sigma / \kappa)}
  \item \eqn{\kappa}: the shrinking factor
  \item \eqn{\Omega}: pseudo sum of squares, \eqn{\Sigma \sim \text{InvWishart}(\nu, \Omega)}
  \item \eqn{\nu}: pseudo degrees of freedom
  \item \eqn{\rho}: correlation degree from previous stick-breaking weights
}
}
\references{
Matthew A. Taddy (2012) Autoregressive Mixture Models for Dynamic Spatial Poisson Processes:
 Application to Tracking Intensity of Violent Crime, Journal of the American Statistical Association, 
 105:492, 1403-1417, DOI: 10.1198/jasa.2010.ap09655
}
