% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dp_normal_deval_conditional}
\alias{dp_normal_deval_conditional}
\title{Eval Point Conditional density}
\usage{
dp_normal_deval_conditional(model, xnew, eval_dims, condition_dims,
  condition_values, nparticles = 50L)
}
\description{
Bla Bla
}
