% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dp_normal_mix}
\alias{dp_normal_mix}
\title{Dirichlet Process Normal Mixture Kernel Density Estimation}
\usage{
dp_normal_mix(x, N, alpha, lambda, kappa, nu, Omega)
}
\arguments{
\item{x}{a matrix of observations}

\item{alpha}{concentration parameter of Dirichlett process}
}
\description{
Bla Bla
}
