% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DynamicParticle}
\alias{DynamicParticle}
\title{DynamicParticle class}
\description{
This class is used inside C++ only and can't be called from R.
It defines a particle, which carries sufficient statistic for the  
distribution of a mixture of normals with dynamic stick-breaking weights'.
The sufficient statistics are:
\itemize{
  \item \eqn{M}: total number of clusters in current period
  \item \eqn{M_0}: total number of clusters in previous period
  \item \eqn{(c_j)_{j=1}^M}: counts of observations in all periods
  \item \eqn{(c^t_j)_{j=1}^M}: counts of observations in current period
  \item \eqn{(w_{j})_{j=1}^M}: stick-breaking weights of each cluster
  \item \eqn{(s_j)_{j=1}^M}: d-dim vector sum of each cluster
  \item \eqn{(SS_j)_{j=1}^M}: dxd-matrix of sum-of-squares of each cluster
}
}
\references{
Matthew A. Taddy (2012) Autoregressive Mixture Models for Dynamic Spatial Poisson Processes:
 Application to Tracking Intensity of Violent Crime, Journal of the American Statistical Association, 
 105:492, 1403-1417, DOI: 10.1198/jasa.2010.ap09655
}
