% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{firstorder_poisson_dlm}
\alias{firstorder_poisson_dlm}
\title{Poisson DLM Random Walk}
\usage{
firstorder_poisson_dlm(counts, m0, C0, nu0, xi0, delta, nparticles = 100)
}
\arguments{
\item{counts}{numeric vector with counts}

\item{m0}{numeric starting point for state variable mean}

\item{C0}{numeric>0 starting point for state variable variance}

\item{nu0}{numeric>0 starting degrees of freedom (pseudocounts) for random walk variance}

\item{xi0}{numeric>0 starting sum of squares for random walk variance}

\item{delta}{0<numeric<=1 correlation parameter}

\item{nparticles}{number of particles for the filter}
}
\description{
Simpel function for filtering with a random walk model
}
