% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iotest_ddpn_hyperparam}
\alias{iotest_ddpn_hyperparam}
\title{Convert a DDPNHyperParam R list to C++ and back}
\usage{
iotest_ddpn_hyperparam(L)
}
\description{
This function is used for unit testing that the process of creating DDPNHyperParam 
in C++ from an R-list and listing them back works correctly. Not be used for other purposes.
}
