% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{list_ddpn_hyperparam}
\alias{list_ddpn_hyperparam}
\alias{list_ddpn_hyperparam}
\title{From DDPNHyperParam in C++ to List in R of class DDPNHyperParam}
\description{
This class is used inside C++ only and can't be called from R.
It reads a list of class DDPNHyperParam from R a constructs the corresponding object in C++

This class is used inside C++ only and can't be called from R.
It reads a list of class DDPN from R a constructs the corresponding object in C++
}
