% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{list_dynamic_particle}
\alias{list_dynamic_particle}
\title{From DynamicParticle in C++ to list in R}
\description{
This class is used inside C++ only and can't be called from R.
It reads a list of class DynamicParicle from R a constructs the corresponding object in C++
}
