\name{pldensity-package}
\alias{pldensity-package}
\alias{pldensity}
\docType{package}
\title{
\packageTitle{pldensity}
}
\description{
\packageDescription{pldensity}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pldensity}
\packageIndices{pldensity}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{pldensity}

Maintainer: \packageMaintainer{pldensity}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
