% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_ddpn}
\alias{read_ddpn}
\title{From List in R to DDPN in C++}
\description{
This class is used inside C++ only and can't be called from R.
It reads a list of class DDPN from R a constructs the corresponding object in C++
}
