% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_ddpn_hyperparam}
\alias{read_ddpn_hyperparam}
\title{From List in R to DDPNHyperParam in C++}
\description{
This class is used inside C++ only and can't be called from R.
It reads a list of class DDPNHyperParam from R a constructs the corresponding object in C++
}
