% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_dynamic_particle}
\alias{read_dynamic_particle}
\title{From List in R to DynamicParticle in C++}
\description{
This class is used inside C++ only and can't be called from R.
It reads a list of class DynamicParicle from R a constructs the corresponding object in C++
}
