% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_plot.R
\name{spatial_plot}
\alias{spatial_plot}
\title{Spatial plots}
\usage{
spatial_plot(model, newx = NULL, view_center = NULL, zoom = 11,
  polygons = FALSE, contour_levels = 250, resol = 25, nparticles = 50)
}
