% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_particle}
\alias{update_particle}
\title{Particle state update}
\description{
Updates the statistics of a particle after seeing data point \code{xnew}. An allocation
of \code{xnew} takes place based on the cluster likelihood.
}
\references{
Matthew A. Taddy (2012) Autoregressive Mixture Models for Dynamic Spatial Poisson Processes:
 Application to Tracking Intensity of Violent Crime, Journal of the American Statistical Association,
 105:492, 1403-1417, DOI: 10.1198/jasa.2010.ap09655
}
